requirements.txt
torch
torchvision
pytorch-lightning
opencv-python
matplotlib
seaborn
scikit-learn
numpy
PyYAML
Pillow
timm
omegaconf
geffnet
pydantic
pycocotools
effdet
albumentations
ensemble-boxes
onnx
onnxruntime
beartype
pre-commit